variant: rename `MessageVariant` to `ReplyVariant`
For more consistency with the rest of the crate, renames the
`MessageVariant` enum to `ReplyVariant`.
