+++ fuzz/fuzz_targets/fuzz_build_message.rs
@@ -2,13 +2,13 @@
 
 use libfuzzer_sys::fuzz_target;
 
-use ebds::{MessageVariant, len};
+use ebds::{ReplyVariant, len};
 
 fuzz_target!(|data: &[u8]| {
     // Omnibus Reply's length is the minimum for any reply message
     if data.len() < len::OMNIBUS_REPLY {return}
 
-    let msg = match MessageVariant::from_buf(data) {
+    let msg = match ReplyVariant::from_buf(data) {
         Ok(msg) => msg,
         Err(_err) => return,
     };
